Eliminate duplication between implementations of ConfigurationSupportInternal
